refactor: address PR #155 review feedback from amito
- Extract estimation logic from config_finder.py to a new
  src/planner/recommendation/estimator.py module, separating roofline
  estimation from capacity planning/scoring (A1)
- Consolidate 3x duplicated stdout/stderr suppression into a shared
  _suppress_noisy_output() context manager (A3)
- Change blocking async route handlers to sync def so FastAPI auto-runs
  them in a thread pool instead of blocking the event loop (A2)
- Move inline import to module level in benchmarks.py (A4)
- Reorder save_benchmarks() to validate data before opening DB
  connection for faster failure and shorter-lived connections (A5)
- Add explicit conn.rollback() on insert failure in
  save_benchmarks() (A6)
- Add unit tests for save_benchmarks() rollback and validation
  ordering
